requirements.txt
dukpy==0.5.0
numpy==2.2.6
pybind11==3.0.1
PySDL2==0.9.17
pysdl2-dll==2.32.0
skia-python==87.9
